+++ NEWS.md
@@ -1,9 +1,31 @@
 # Spectra 1.17
 
-## Change in 1.17.6
+
+## Change in 1.17.10
 
 - Accept labels argument as a `list` instead of a `character` in the 
   plotting functions. 
+
+## Change in 1.17.9
+
+- Allow parameter `msLevel. = integer()` for `filterRt()` to filter spectra of
+  **all** MS levels. This was `msLevel. = uniqueMsLevels()`, which, depending
+  on the backend, can be computationally intense. Add related unit tests to
+  the unit test suite.
+
+## Change in 1.17.8
+
+- Add parameter `return.type` to `peaksData()`.
+
+## Change in 1.17.7
+
+- Add the `spectraVariableMapping<-` generic method.
+
+## Change in 1.17.6
+
+- Add new `fillCoreSpectraVariables()` function that allows to add eventually
+  missing *core* spectra variables (with the correct data type) to a data frame.
+
 
 ## Change in 1.17.5